Support data references made by 16 bit code
The analyzer assumes DS and CS to be the segment base address and using
this information maps 16 bit ?s:data references to the virtual address
space.
The printer performs the same virtual address conversion and looks up
labels in the labelTypes map to replace literals with labels in 16 bit
code.
Note that the labels are not emitted by the printer as GCC cannot handle
them. The related code snippet is commented out.