Add support to load map file
le_disasm is not perfect, but with some aid from the users in the form
of manual symbol definitions the output accuracy of the disassembler can
be exponentially improved. The change set allows to load a map file
specified as an input argument to the tool and processes the inputs.